Split out versioning.
The windows version is used for the unix versions,
so can't have its computation only run when running on windows.
